Refactor Elasticsearch integration
Comprehensive refactoring of Elasticsearch integration:
- Removed `TC.CloudGames.Games.Search` project and migrated functionality to `Infrastructure`.
- Introduced `ElasticSearchOptions` and `ElasticsearchClientProvider` for centralized configuration and client management.
- Enhanced dependency injection with `AddElasticSearch` for seamless service registration.
- Optimized `ElasticGameSearchService` for Elasticsearch 9.x with advanced search, bulk indexing, and aggregation improvements.
- Updated endpoints with improved logging (using emojis) and error handling.
- Removed redundant integration events and deprecated files.
- Centralized Elasticsearch configuration with environment-specific support.
- Overhauled documentation with new guides and examples.
- Updated project files and removed unused example/test folders.

These changes simplify the codebase, improve maintainability, and ensure compatibility with Elasticsearch Cloud Serverless and local development.
