+++ src/Adapters/Inbound/TC.CloudGames.Games.Api/appsettings.Development.json
@@ -50,7 +50,7 @@
     "Schema": "public",
     "ConnectionTimeout": 30
   },
-  "Elasticsearch__": {
+  "Elasticsearch_Local": {
     "Host": "localhost",
     "Port": 9200,
     "Username": "elastic",
@@ -58,8 +58,8 @@
     "IndexPrefix": "games"
   },
   "Elasticsearch": {
-    "EndpointUrl": "https://my-elasticsearch-project-f61917.es.eastus.azure.elastic.cloud:443",
-    "ApiKey": "cTVic2w1a0JRTVgyQzZtWnMxemM6dG1qb2VIWmQ4bXhsY04zOER5Y2d1dw==",
+    "EndpointUrl": "<elasticsearch-cloud-url>",
+    "ApiKey": "<ApiKey>",
     "ProjectId": "f619171f1605489b90e4501ce1940394",
     "IndexPrefix": "search-xn8c",
     "RequestTimeoutSeconds": 30,
