+++ src/Adapters/Inbound/TC.CloudGames.Games.Api/appsettings.elasticsearch.example.json
