+++ src/Adapters/Outbound/TC.CloudGames.Games.Infrastructure/DependencyInjection.cs
@@ -1,4 +1,7 @@
-﻿using TC.CloudGames.SharedKernel.Infrastructure.Snapshots.Users;
+﻿using Microsoft.Extensions.Configuration;
+using TC.CloudGames.Games.Infrastructure.Elasticsearch;
+using TC.CloudGames.SharedKernel.Infrastructure.Elasticsearch;
+using TC.CloudGames.SharedKernel.Infrastructure.Snapshots.Users;
 
 namespace TC.CloudGames.Games.Infrastructure
 {
@@ -16,6 +19,21 @@ namespace TC.CloudGames.Games.Infrastructure
             services.AddSingleton<ITokenProvider, TokenProvider>();
             services.AddScoped<IUserContext, UserContext>();
             services.AddScoped<IUserSnapshotStore, UserSnapshotStore>();
+
+            return services;
+        }
+
+        public static IServiceCollection AddElasticSearch(this IServiceCollection services, IConfiguration configuration)
+        {
+            // Configure Elasticsearch options
+            services.Configure<ElasticSearchOptions>(configuration.GetSection("Elasticsearch"));
+
+            // Register Elasticsearch client provider and client
+            services.AddSingleton<IElasticsearchClientProvider, ElasticsearchClientProvider>();
+            services.AddSingleton(sp => sp.GetRequiredService<IElasticsearchClientProvider>().Client);
+
+            // Register search services
+            services.AddScoped<IGameSearchService, ElasticGameSearchService>();
 
             return services;
         }
