+++ src/Adapters/Outbound/TC.CloudGames.Games.Search/ElasticsearchServiceCollectionExtensions.cs
@@ -1,0 +1,20 @@
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using TC.CloudGames.SharedKernel.Infrastructure.Elasticsearch;
+
+namespace TC.CloudGames.Games.Search;
+
+/// <summary>
+/// Extension methods for configuring Elasticsearch services.
+/// </summary>
+public static class ElasticsearchServiceCollectionExtensions
+{
+    /// <summary>
+    /// Adds Elasticsearch services to the dependency injection container.
+    /// Configures both the client provider and search services.
+    /// </summary>
+    /// <param name="services">The service collection</param>
+    /// <param name="configuration">The configuration</param>
+    /// <returns>The service collection for chaining</returns>
+    
+}
